Extract module pyk.prelude.k
